Batch serial console input bytes in CLI (#104)
In serial task, batch reads from tokio::io::stdin, and send that as one
Binary Websockets message instead of one u8 at a time. This fixes the
dreaded "[6;14R" junk that was seen in serial sessions when the propolis
server was zoned.

I suspect this is due to some timeout logic somewhere in the giant stack
of software that only considers ESC [ a [Control Sequence Introducer][1]
if it's followed by the other bytes "quickly" enough.

There were *many* red herrings during this dive, including the rewriting
of the serial task to use a channel instead of for stdin.read as a
tokio::select branch - reading from tokio::io::stdin isn't cancel safe
so it shouldn't be on one of those branches.

Multiple people have looked at this, and so did I, and going into it
with the assumption that a byte was being lost somewhere caused me to
read a lot of the plumbing code surrounding the UART logic. This commit
also adds some tests for copy_and_consume because I noticed there
weren't any.

[1]: https://en.wikipedia.org/wiki/ANSI_escape_code#CSI_(Control_Sequence_Introducer)_sequences